refactor: #51 도메인별 패키지 구조 개선
